+++ pkg/client_option.go
@@ -1,0 +1,89 @@
+package pkg
+
+import (
+	"errors"
+	"fmt"
+	"net/http"
+	"time"
+)
+
+// ClientOption is used to set Client settings.
+type ClientOption func(c *Client) error
+
+// NewClient initializes the cache HTTP middleware client with the given
+// options.
+func NewClient(opts ...ClientOption) (*Client, error) {
+	c := &Client{}
+
+	for _, opt := range opts {
+		if err := opt(c); err != nil {
+			return nil, err
+		}
+	}
+
+	if c.adapter == nil {
+		return nil, errors.New("cache client adapter is not set")
+	}
+	if int64(c.ttl) < 1 {
+		return nil, errors.New("cache client ttl is not set")
+	}
+	if c.methods == nil {
+		c.methods = []string{http.MethodGet}
+	}
+
+	return c, nil
+}
+
+// ClientWithAdapter sets the adapter type for the HTTP cache
+// middleware client.
+func ClientWithAdapter(a Adapter) ClientOption {
+	return func(c *Client) error {
+		c.adapter = a
+		return nil
+	}
+}
+
+// ClientWithTTL sets how long each response is going to be cached.
+func ClientWithTTL(ttl time.Duration) ClientOption {
+	return func(c *Client) error {
+		if int64(ttl) < 1 {
+			return fmt.Errorf("cache client ttl %v is invalid", ttl)
+		}
+
+		c.ttl = ttl
+
+		return nil
+	}
+}
+
+// ClientWithRefreshKey sets the parameter key used to free a request
+// cached response. Optional setting.
+func ClientWithRefreshKey(refreshKey string) ClientOption {
+	return func(c *Client) error {
+		c.refreshKey = refreshKey
+		return nil
+	}
+}
+
+// ClientWithMethods sets the acceptable HTTP methods to be cached.
+// Optional setting. If not set, default is "GET".
+func ClientWithMethods(methods []string) ClientOption {
+	return func(c *Client) error {
+		for _, method := range methods {
+			if method != http.MethodGet && method != http.MethodPost {
+				return fmt.Errorf("invalid method %s", method)
+			}
+		}
+		c.methods = methods
+		return nil
+	}
+}
+
+// ClientWithExpiresHeader enables middleware to add an Expires header to responses.
+// Optional setting. If not set, default is false.
+func ClientWithExpiresHeader() ClientOption {
+	return func(c *Client) error {
+		c.writeExpiresHeader = true
+		return nil
+	}
+}
